% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tree_geolocations.R
\name{calc_tree_geolocations}
\alias{calc_tree_geolocations}
\title{calculate NEON woody vegetation structure individual geolocations.
more about the geoNEON package here:
https://github.com/NEONScience/NEON-geolocation/tree/master/geoNEON}
\usage{
calc_tree_geolocations(data = vst, dataProd = "vst_mappingandtagging")
}
\value{
A list of dataframe
}
\description{
calculate NEON woody vegetation structure individual geolocations.
more about the geoNEON package here:
https://github.com/NEONScience/NEON-geolocation/tree/master/geoNEON
}
\examples{
calc_tree_geolocations(data = vst, dataProd = "vst_mappingandtagging")



}
\seealso{
[geoNEON::def.calc.geo.os()] which this function wraps
}
