% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_coords_itc.R
\name{retrieve_coords_itc}
\alias{retrieve_coords_itc}
\title{get individualID coordinates}
\usage{
retrieve_coords_itc(dat)
}
\value{
a dataframe
}
\description{
function to calculate the coordinates for each
individual tree in the vegetation structure survay
}
\examples{
retrieve_VST_data("OSBS")
retrieve_VST_data("all", 2017, 2019)


}
\seealso{
[geoNEON::loadByProduct()] which this function wraps.
}
