% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_VST_data.R
\name{retrieve_VST_data}
\alias{retrieve_VST_data}
\title{retrieve vegetation structure data  from NEON}
\usage{
retrieve_VST_data(site = "all", start, enddate)
}
\value{
A list of dataframe
}
\description{
retrieve vegetation structure data  from NEON
}
\examples{
retrieve_VST_data("OSBS")
retrieve_VST_data("all", 2017, 2019)


}
\seealso{
[neonUtilities::loadByProduct()] which this function wraps.
}
